d_to_g {MAd}	R Documentation

Standardized Mean Difference (d) Statistic to Unbiased Standardized Mean Difference (g)

Description

Eliminates the small upward bias of d to provide an unbiased estimate of the population effect size parameter (g). This procedure will compute g for a single value of d.

Usage

d_to_g(d, var.d, n.1, n.2)

Arguments

d
Standardized mean difference statistic (d has a slight bias).

var.d
Variance of d.

n.1
Sample size of treatment group.

n.2
Sample size of comparison group.

Value

g
Unbiased Standardized mean difference statistic.

var.g
Variance of g.

Author(s)

AC Del Re & William T. Hoyt

Maintainer: AC Del Re acdelre@gmail.com

References

Borenstein (2009). Effect sizes for continuous data. In H. Cooper, L. V. Hedges, & J. C. Valentine (Eds.), The handbook of research synthesis and meta analysis (pp. 279-293). New York: Russell Sage Foundation.

[Package MAd version 0.8-2.1 Index]