dat.hoyt {MAd}	R Documentation

Subset of simulated data to demonstrate aggregation for dependent effect sizes

Description

Subset of simulated data to demonstrate aggregation for dependent effect sizes. This data comes from: Hoyt, W. T., & Del Re, A. C. (2013). Comparison of methods for aggregating dependent effect sizes in meta-analysis. Manuscript submitted for publication.

[Package MAd version 0.8-2.1 Index]