mean_to_d2 {MAd}	R Documentation

Means with Pooled SD to Standardized Mean Difference

Description

Converts raw mean scores with pooled standard deviation from the primary study to a standardized mean difference (d).

Usage

mean_to_d2(m.1, m.2, s.pooled, n.1, n.2)

Arguments

m.1
Mean of group one.

m.2
Mean of group two.

s.pooled
Pooled standard deviation.

n.1
Sample size of group one.

n.2
Sample size of group two.

Value

d
Standardized mean difference (d).

var.d
Variance of d.

Author(s)

AC Del Re & William T. Hoyt

Maintainer: AC Del Re acdelre@gmail.com

References

Borenstein (2009). Effect sizes for continuous data. In H. Cooper, L. V. Hedges, & J. C. Valentine (Eds.), The handbook of research synthesis and meta analysis (pp. 279-293). New York: Russell Sage Foundation.

See Also

d_to_g, mean_to_d, t_to_d, f_to_d, p_to_d1, p_to_d2, ancova_to_d1, ancova_to_d2, tt.ancova_to_d, f.ancova_to_d, r_to_d, p.ancova_to_d1, p.ancova_to_d2, lor_to_d, prop_to_or, prop_to_d, r_from_chi, r_from_d, r_from_d1, r_from_t

[Package MAd version 0.8-2.1 Index]