p.ancova_to_d1 {MAd}	R Documentation

One-tailed p-value from ANCOVA to Standardized Mean Difference (d)

Description

Converts a one-tailed p-value from ANCOVA reported in the primary study to a standardized mean difference (d)

Usage

p.ancova_to_d1(p, n.1, n.2, R, q)

Arguments

p
One-tailed p-value reported in primary study.

n.1
Treatment group sample size.

n.2
Comparison group sample size.

R
Covariate outcome correlation or multiple correlation.

q
number of covariates.

Value

d
Standardized mean difference (d).

var_d
Variance of d.

Author(s)

AC Del Re & William T. Hoyt

Maintainer: AC Del Re acdelre@gmail.com

References

Borenstein (2009). Effect sizes for continuous data. In H. Cooper, L. V. Hedges, & J. C. Valentine (Eds.), The handbook of research synthesis and meta analysis (pp. 279-293). New York: Russell Sage Foundation.

See Also

d_to_g, mean_to_d, mean_to_d2, t_to_d, f_to_d, p_to_d1, p_to_d2, ancova_to_d1, ancova_to_d2, tt.ancova_to_d, f.ancova_to_d, r_to_d, p.ancova_to_d2, lor_to_d, prop_to_or, prop_to_d, r_from_chi, r_from_d, r_from_d1, r_from_t

[Package MAd version 0.8-2.1 Index]